Refs #51 (0.5) Added intentions and reasons while programming on #28 Refs #28 (4.75) Fixed expanding, now only expands once SelfClassifier added: isScouting(), isHighOnMinerals(), isHighOnGas(), isHighOnResources() Commander: Tested onUpgradeSoonFinish -> attack On expansion low on minerals -> expand High on resources -> transitions to the next phase orderJoin() function added for joining frontal attack with an allied squad. Ping now works for attacks, changed logic in AttackSquad, not using createGoal() anymore. AttackSquad now never automatically joins an allied squad.

+++ BATS/code/BATSModule/src/GameTime.cpp
@@ -47,5 +47,5 @@ double GameTime::convertFramesToSeconds(int frames) {
 }
 
 int GameTime::convertSecondsToFrames(double seconds) {
-	return seconds * FRAMES_PER_SECOND;
+	return static_cast<int>(seconds * FRAMES_PER_SECOND);
 }
